refactor: fix interface and exports
